requirements.txt
pandas
numpy
joblib
python-binance
ta
scikit-learn
streamlit